requirements.txt
unstructured[pdf]
fitz
PyMuPDF
langchain-groq
langchain-openai
python-dotenv